requirements.txt
asgiref==3.10.0
Django==5.2.8
python-decouple==3.8
sqlparse==0.5.3
pymongo